+++ services/web/server/src/simcore_service_webserver/users/_notifications_handlers.py
@@ -20,6 +20,7 @@ from ..redis import get_redis_user_notifications_client
 from ..security.decorators import permission_required
 from ..utils_aiohttp import envelope_json_response
 from . import _users_service
+from ._common.schemas import UsersRequestContext
 from ._notifications import (
     MAX_NOTIFICATIONS_FOR_USER_TO_KEEP,
     MAX_NOTIFICATIONS_FOR_USER_TO_SHOW,
@@ -28,7 +29,6 @@ from ._notifications import (
     UserNotificationPatch,
     get_notification_key,
 )
-from .common._schemas import UsersRequestContext
 
 _logger = logging.getLogger(__name__)
